SAPPORO-web/requirements.dev.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.2
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-bootstrap4==0.0.7
django-debug-toolbar==1.11
factory-boy==2.11.1
Faker==1.0.2
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==2.1.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.24.1
Werkzeug==0.14.1
